requirements.txt
Flask
google-generativeai
moviepy==1.0.3
SpeechRecognition
pydub
werkzeug
